requirements.txt
strands-agents>=0.1.0
strands-agents-tools>=0.1.0
pyyaml>=6.0
python-dotenv>=1.0.0
requests>=2.31.0
rich>=13.0.0
